+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * In general follow (https://docs.npmjs.com/getting-started/semantic-versioning) versioning.
 ## <next>
+* Arrows for reordering tree elements (`showOrderingArrows` and `onOrderButtonClick` props)
 * Added `@opuscapita/react-perfect-scrollbar` around tree
 * Made some changes on how nodes are expanded. Now it's not necessary to provide external `onExpand` callback in most of the cases.